refactor: scrim 태그 분리
