+++ src/components/portal/BottomSheet.tsx
@@ -3,6 +3,7 @@ import styled from '@emotion/styled';
 import { m, Variants } from 'framer-motion';
 
 import PortalWrapper from './PortalWrapper';
+import StyledMotionScrim from './StyledMotionScrim';
 
 import { defaultEasing, defaultFadeInVariants } from '@/constants/motions';
 import usePreventScroll from '@/hooks/scroll/usePreventScroll';
@@ -35,19 +36,6 @@ const BottomSheet = ({ setToClose, children, isShowing }: Props) => {
 };
 
 export default BottomSheet;
-
-const StyledMotionScrim = styled(m.div)`
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 900;
-
-  width: 100vw;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.3);
-
-  overflow: hidden;
-`;
 
 const StyledMotionContent = styled(m.div)`
   position: absolute;
